refactor: 模块扩展参数目录重构
