+++ src/utils/module-extends-options/config.ts
@@ -1,0 +1,7 @@
+import type { AppRouterOptions } from './type';
+
+export function getConfigExtendsOptions(_options: AppRouterOptions) {
+  return { };
+}
+
+export type ConfigExtendsOptions = ReturnType<typeof getConfigExtendsOptions>;
