+++ src/utils/module-extends-options/controller.ts
@@ -1,0 +1,7 @@
+import type { AppRouterOptions } from './type';
+
+export function getControllerExtendsOptions(_options: AppRouterOptions) {
+  return { };
+}
+
+export type ControllerExtendsOptions = ReturnType<typeof getControllerExtendsOptions>;
