+++ src/utils/module-extends-options/extend.ts
@@ -1,0 +1,7 @@
+import type { AppRouterOptions } from './type';
+
+export function getExtendExtendsOptions(_options: AppRouterOptions) {
+  return { };
+}
+
+export type ExtendExtendsOptions = ReturnType<typeof getExtendExtendsOptions>;
