+++ src/utils/module-extends-options/router-schema.ts
@@ -1,0 +1,7 @@
+import type { AppRouterOptions } from './type';
+
+export function getRouterSchemaExtendsOptions(_options: AppRouterOptions) {
+  return { };
+}
+
+export type RouterSchemaExtendsOptions = ReturnType<typeof getRouterSchemaExtendsOptions>;
