+++ src/utils/module-extends-options/router.ts
@@ -1,0 +1,8 @@
+import type KoaRouter from '@koa/router';
+import type { AppRouterOptions } from './type';
+
+export function getRouterExtendsOptions(_options: AppRouterOptions & { globalRouter: KoaRouter }) {
+  return { };
+}
+
+export type RouterExtendsOptions = ReturnType<typeof getRouterExtendsOptions>;
