+++ src/utils/module-extends-options/service.ts
@@ -1,0 +1,7 @@
+import type { AppRouterOptions } from './type';
+
+export function getServiceExtendsOptions(_options: AppRouterOptions) {
+  return { };
+}
+
+export type ServiceExtendsOptions = ReturnType<typeof getServiceExtendsOptions>;
