+++ src/utils/module-extends-options/type.ts
@@ -1,0 +1,3 @@
+import type { TeeKoa } from '../../types';
+
+export type AppRouterOptions = TeeKoa.AppOptions & TeeKoa.RouterOptions;
